EngineSetup
Implement EngineSetup class to keep bloat for setting up a game from main class.
